email_dashboard_project/requirements.txt
streamlit==1.37.1
SQLAlchemy==2.0.32
schedule==1.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.37.1
loguru==0.7.2
streamlit-datetime-range-picker==0.0.1

